+++ Final/code/include/utils.hpp
@@ -4,11 +4,11 @@
 #include <vecmath.h>
 #include "ray.hpp"
 
-int sgn(float f);
+int sgn(double f);
 
-float calcT(const Ray &r, const Vector3f &p);
+double calcT(const Ray &r, const Vector3f &p);
 
-float calcReLU(const float &f);
+double calcReLU(const double &f);
 
 void printVector3f(const Vector3f &f);
